@@ -2,10 +2,11 @@
   "name": "@cbcruk/release-cli",
   "version": "1.0.0",
   "bin": {
-    "release": "bin/release.js"
+    "release": "dist/cli.js"
   },
   "scripts": {
-    "dev": ""
+    "build": "tsup",
+    "dev": "tsup --watch"
   },
   "dependencies": {
     "@fluent/bundle": "^0.16.0",
@@ -18,6 +19,22 @@
     "semver": "^7.3.2"
   },
   "devDependencies": {
-    "eslint": "^7.5.0"
+    "@cli-/tsconfig": "*",
+    "@types/lodash": "latest",
+    "@types/node": "^16.11.13",
+    "@types/prompts": "latest",
+    "@types/semver": "^7.3.10",
+    "eslint": "^7.32.0",
+    "tsup": "^5.11.6",
+    "typescript": "^4.5.4"
+  },
+  "tsup": {
+    "entry": [
+      "src/cli.ts"
+    ],
+    "splitting": false,
+    "sourcemap": true,
+    "clean": true,
+    "minify": true
   }
 }
